feat(wasm): handle pack issues
