+++ packages/pack-shared/src/magicIdentifier.ts
