+++ packages/pack/package.json
@@ -38,6 +38,7 @@
   "license": "MIT",
   "dependencies": {
     "@utoo/style-loader": "^1.0.0",
+    "@utoo/pack-shared": "^0.0.4",
     "react-refresh": "^0.12.0",
     "@babel/code-frame": "7.22.5",
     "@swc/helpers": "0.5.15",
